backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
